+++ src/types/typings.ts
@@ -75,6 +75,7 @@ export interface HashResult {
 
 export interface LocalHashObj extends HashResult {
   failedLoginAttempts?: number;
+  lastFailedLogin?: number;
   lockedUntil?: number;
 }